Basic Add Student component working
