+++ src/App.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import requiresAuth from './auth/requiresAuth';
 import { Route } from 'react-router-dom';
 import { LoginView, HeaderView, SponsorChildView, BoardView, StudentView } from './views';
-import { Onboarding } from './components'
+import { Onboarding, AddStudent } from './components'
 
 
 import './App.css';
@@ -16,6 +16,7 @@ class App extends Component {
         <Route path='/sponsor' component={SponsorChildView} />
         <Route path='/board' component={BoardView} />
         <Route path='/student/:id' render={props => <StudentView {...props} />} />
+        <Route path='/add' component={AddStudent} />
       </div>
     );
   }
