+++ src/views/index.js
@@ -4,5 +4,4 @@ export { default as HeaderView } from './HeaderView';
 export { default as BoardView } from './BoardView';
 export { default as SponsorChildView } from './SponsorChildView';
 export { default as StudentView } from './StudentView';
-export { default as AddStudent } from './AddStudent';
